% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{check_code_in_rstudio}
\alias{check_code_in_rstudio}
\title{Check whether the code crash RStudio or not.}
\usage{
check_code_in_rstudio(code)
}
\arguments{
\item{code}{the code you want to test in RStudio}
}
\description{
\code{check_code-in_rstudio} checks whether the code crash RStudio or not.
}
\examples{
check_code_in_rstudio("1")

}
