% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{check_in_rstudio}
\alias{check_in_rstudio}
\title{Check the code in RStudio.}
\usage{
check_in_rstudio(code, time = 30)
}
\arguments{
\item{code}{the code you want to test in RStudio}

\item{time}{the time for the testing}
}
\value{
whether or not the code runs in RStudio successfully.
}
\description{
\code{check_in_rstudio} checks the code in RStudio.
}
\examples{
\dontrun{
check_in_rstudio("1")
}

}
