% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{check_rstudio}
\alias{check_rstudio}
\title{Check whether RStudio is available or not.}
\usage{
check_rstudio()
}
\description{
\code{check_rstudio} checks whether RStudio is available or not.
}
\examples{
\dontrun{
check_rstudio()
}

}
