% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{detailed_check_in_rstudio}
\alias{detailed_check_in_rstudio}
\title{Check the code in RStudio.}
\usage{
detailed_check_in_rstudio(code, time = 20)
}
\arguments{
\item{code}{the code you want to test in RStudio}

\item{time}{the time for the testing}
}
\value{
a list,
    the component crashed means whether rsession is crashed,
    the component finished means whether the code runs to the end successfully,
    the component errmsg is the error message of the code (if any).
}
\description{
\code{detailed_check_in_rstudio} checks the code in RStudio.
}
\examples{
\dontrun{
detailed_check_in_rstudio("1")
}

}
