% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{inject_into_rstudio}
\alias{inject_into_rstudio}
\title{Start a new RStudio process and "inject" some codes into it.}
\usage{
inject_into_rstudio(code)
}
\arguments{
\item{code}{the code you want to inject into RStudio}
}
\description{
\code{inject_into_rstudio} start a new RStudio process and "inject" some codes into it.
}
\examples{
inject_into_rstudio("1")

}
