% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setup}
\alias{setup}
\title{Do initial setup for the TestWithRStudio package.}
\usage{
setup()
}
\description{
\code{setup} does the initial setup for the TestWithRStudio package.
}
\examples{
crash <- setup()

}
