% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{check_code_in_rstudio}
\alias{check_code_in_rstudio}
\title{Check whether the code crash RStudio or not.}
\usage{
check_code_in_rstudio(code, time = 10)
}
\arguments{
\item{code}{the code you want to test in RStudio}

\item{time}{the time for the testing}
}
\description{
\code{check_code_in_rstudio} checks whether the code crash RStudio or not.
}
\examples{
stopifnot(check_code_in_rstudio("1"))
stopifnot(!check_code_in_rstudio("q()"))
stopifnot(!check_code_in_rstudio("library(TestWithRStudio); crash()"))

}
